Move Country class to the appconfig package
