+++ app/src/main/java/org/simple/clinic/selectcountry/SelectCountryEvent.kt
@@ -1,5 +1,7 @@
 package org.simple.clinic.selectcountry
 
+import org.simple.clinic.appconfig.Country
+
 sealed class SelectCountryEvent
 
 data class ManifestFetched(val countries: List<Country>) : SelectCountryEvent()
