+++ app/src/test/java/org/simple/clinic/selectcountry/SelectCountryUpdateTest.kt
@@ -7,6 +7,7 @@ import com.spotify.mobius.test.NextMatchers.hasNoModel
 import com.spotify.mobius.test.UpdateSpec
 import com.spotify.mobius.test.UpdateSpec.assertThatNext
 import org.junit.Test
+import org.simple.clinic.appconfig.Country
 import java.net.URI
 
 class SelectCountryUpdateTest {
